kit: extract CommonSshOptions into ssh_options.rs
Extract CommonSshOptions and SshConnectionOptions into a dedicated
cross-platform module. These types were embedded in ssh.rs alongside
Linux-specific container SSH logic, making them unavailable to other
backends.
