+++ crates/kit/src/lib.rs
@@ -2,6 +2,7 @@
 
 pub mod cpio;
 pub mod qemu_img;
+pub mod ssh_options;
 pub mod xml_utils;
 
 // Linux-only modules
